// src/app/@api/model/profileDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ProfileCompanyDetailsDto } from './profileCompanyDetailsDto';
import { ProfileUserDto } from './profileUserDto';
import { ProfileUserProfileDto } from './profileUserProfileDto';


export interface ProfileDto { 
    company?: ProfileCompanyDetailsDto;
    user?: ProfileUserDto;
    userProfile?: ProfileUserProfileDto;
}